[Heartbeat] Mock out background loop in fixture
Mock out `Core._loop` in `cogHeartbeat` fixture to avoid the background
loop being run. This allows us to say that the background loop does not
run upon cog instantiation via the fixture, rather than cancelling the
background task knowing that the background loop might have run for a
bit. The latter poses uncertainty and is not preferable.