[VxTwit] Add `ddinstagram.com` replacements
This commit adds event listeners to replace instagram.com links
with ddinstagram.com. It is using the same toggle as VxTwit.

Since instagram.com embeds don't have videos or photos, we will need to
replace all embeds. As such, the `valid` method was modified to not
check for videos in embeds: this is deferred to a later check for the
VxTwitter case.

This commit doesn't change the cog name; this will be done in a follow
up.