[Gatekeep] Add attachment scoring, updated subcommand tree
Scoring for attachments in messages

From observing the behaviour of recent spam messages, this commit adds
the ability to score attachments in messages. There is a new
configurable weight for attachments, called `attachmentweight`. For each
attachment in the message, the score of the message increases by
`attachmentweight`. By default, this is set to 0.

To change the weight for attachments, use the following command or any
of its aliases:

- [p]gatekeep settings attachmentweight <weight>
- [p]gatekeep settings attachments <weight>
- [p]gatekeep settings attachment <weight>
- [p]gatekeep settings aw <weight>

`[p]gatekeep test <message>` is updated to support testing messages with
attachments.

The base [p]gatekeep <subcommand> tree was getting cluttered, especially
with the newly added attachmentweight setting. 4 commands were moved
from the base subcommand tree into their own separate subcommand tree
for a more organized subcommand structure.

This new subcommand tree is aptly named settings, with a few aliases
for convenience:

- [p]gatekeep settings <subcommand>
- [p]gatekeep set <subcommand>
- [p]gatekeep s <subcommand>

The 4 commands that have been moved to this subcommand tree are:

- attachmentweight
- channel
- days
- threshold